Move timer state into GameContext and add client to server logging from GameContext
